requirements.txt
pygame-ce
moderngl
noise
numpy
